# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "API documentation description",
        "version": "1.0.0"
    },
    "paths": {
        "/admin/admin-users": {
            "get": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "Get a paginated list of admin users",
                "operationId": "20e8a3c4187a5e3798452c66882547f1",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of admin users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "John"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "Doe"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "admin@example.com"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "Create a new admin user",
                "operationId": "0ff865b63ed9b402d240c8f0dfba59eb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin user created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/admin/admin-users/{id}": {
            "put": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "Update an existing admin user",
                "operationId": "ac3c8b8af04ef68c971cba3efa5daf6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Admin user ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Updated"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "User"
                                    },
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin user updated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "Delete an admin user",
                "operationId": "de895aaad27b3bcb2ed30e339bf92e0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Admin user ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin user deleted"
                    },
                    "404": {
                        "description": "Admin user not found"
                    }
                }
            }
        },
        "/admin/admin-users/{id}/resend-activation": {
            "post": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "Resend activation email to the admin user",
                "operationId": "00e7e0fc6c856e54447bf04592f83049",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Admin user ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activation email sent"
                    },
                    "409": {
                        "description": "Failed to send activation email"
                    }
                }
            }
        },
        "/admin/admin-users/{id}/impersonal-login": {
            "post": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "Log in as another admin user (impersonal login)",
                "operationId": "030458d5d128e5ec6965c906f5c5b128",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Admin user ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to previous page as impersonated user"
                    }
                }
            }
        },
...
    },
...
}